Prevent screen capture in mnemonic modal
Closes #252
