+++ package.json
@@ -50,6 +50,7 @@
     "expo-linear-gradient": "~12.3.0",
     "expo-local-authentication": "~13.4.1",
     "expo-localization": "~14.3.0",
+    "expo-screen-capture": "~5.3.0",
     "expo-secure-store": "~12.3.1",
     "expo-status-bar": "~1.6.0",
     "expo-web-browser": "~12.3.2",
