+++ src/navigation/RootStackNavigation.tsx
@@ -50,7 +50,7 @@ import NewWalletSuccessScreen from '~/screens/new-wallet/NewWalletSuccessScreen'
 import PinCodeCreationScreen from '~/screens/new-wallet/PinCodeCreationScreen'
 import SelectImportMethodScreen from '~/screens/new-wallet/SelectImportMethodScreen'
 import ProgressHeader from '~/screens/SendReceive/ProgressHeader'
-import SettingsScreen from '~/screens/SettingsScreen'
+import SettingsScreen from '~/screens/Settings/SettingsScreen'
 import SplashScreen from '~/screens/SplashScreen'
 import SwitchWalletScreen from '~/screens/SwitchWallet/SwitchWalletScreen'
 import VerifyMnemonicScreen from '~/screens/VerifyMnemonicScreen'
